Tests from paper, also p-p test.  Test outputs added to .gitignore.
As written, the code will fail the p-p test for sigma_2, the small-variance
dimension.  This is due to a mismatch between the finite test prior (sigma_2
\in [0.1, 0.4]) and the (improper) analysis prior (sigma_2 > 0).  With 100
observations (default in the test), the width of the posterior in sigma_2 is
about sigma_2/10.  For sigma_2 within ~0.04 of the upper test prior boundary,
this means that the posterior will extend past the cutoff of the correct prior,
leading to bias.  But, such sigma_2 are about 10% of the total runs, leading to
failing p-p tests.  For the other variables, the posterior width is a much
smaller fraction of the prior range, so the bias is much smaller.